added table design and created api

+++ src/main/java/com/box/bookstore/controller/AdminpageController.java
@@ -17,6 +17,8 @@ import com.box.bookstore.model.StaffModel;
 import com.box.bookstore.service.DoctorService;
 import com.box.bookstore.service.StaffService;
 
+import jakarta.servlet.http.HttpSession;
+
 @Controller
 public class AdminpageController {
 	
@@ -32,13 +34,23 @@ public class AdminpageController {
 	
 	
 	@GetMapping("/adminuserinterface")
-	public String getAdminUserInterface(Model model) {
+	public String getAdminUserInterface(Model model,HttpSession httpSession) {
+		
+		if(httpSession.getAttribute("validAdmin")==null) {
+			return "adminlogin";
+		}
+		
 		model.addAttribute("indexIndicator","active");
 		return "adminuserinterface";
 	}
 	
 	@GetMapping("/patientlist")
-	public String getAllPatientList(Model model) {
+	public String getAllPatientList(Model model,HttpSession httpSession) {
+		
+		if(httpSession.getAttribute("validAdmin")==null) {
+			return "adminlogin";
+		}
+		
 		try {
 		List<PatientModel> patientlist=patientListApi.getPatientList();
 		model.addAttribute("patient_list", patientlist);
@@ -65,7 +77,12 @@ public class AdminpageController {
 	
 	
 	@GetMapping("/doctorlist")
-	public String getAllDoctorList(Model model) {
+	public String getAllDoctorList(Model model,HttpSession httpSession) {
+		
+		if(httpSession.getAttribute("validAdmin")==null) {
+			return "adminlogin";
+		}
+		
 		List<DoctorModel> doctorModels=doctorService.getAllDoctor();
 		List<DoctorModel> doctorList = null;
 		doctorList=new ArrayList<>();
@@ -83,7 +100,12 @@ public class AdminpageController {
 	
 
 	@GetMapping("/doctorRequestlist")
-	public String getAllDoctorRequestList(Model model) {
+	public String getAllDoctorRequestList(Model model,HttpSession httpSession) {
+		
+		if(httpSession.getAttribute("validAdmin")==null) {
+			return "adminlogin";
+		}
+		
 		List<DoctorModel> doctorModels=doctorService.getAllDoctor();
 		List<DoctorModel> doctorList = null;
 		doctorList=new ArrayList<>();
@@ -102,31 +124,56 @@ public class AdminpageController {
 	}
 	
 	@GetMapping("/deleteRequestDoctor")
-	public String getDeleteRequestDoctor(@RequestParam int id) {
+	public String getDeleteRequestDoctor(@RequestParam int id,HttpSession httpSession) {
+		
+		if(httpSession.getAttribute("validAdmin")==null) {
+			return "adminlogin";
+		}
+		
 		doctorService.deleteDoctor(id);
 		return "redirect:/doctorRequestlist";
 	}
 	
 	@GetMapping("/deleteDoctor")
-	public String getDeleteDoctor(@RequestParam int id) {
+	public String getDeleteDoctor(@RequestParam int id,HttpSession httpSession) {
+		
+		if(httpSession.getAttribute("validAdmin")==null) {
+			return "adminlogin";
+		}
+		
 		doctorService.deleteDoctor(id);
 		return "redirect:/doctorlist";
 	}
 	
 	@GetMapping("/acceptDoctor")
-	public String getAcceptDoctor(@RequestParam int id) {
+	public String getAcceptDoctor(@RequestParam int id,HttpSession httpSession) {
+		
+		if(httpSession.getAttribute("validAdmin")==null) {
+			return "adminlogin";
+		}
+		
 		doctorService.acceptDoctor(id);
 		return "redirect:/doctorRequestlist";
 	}
 	
 	@GetMapping("/addStaff")
-	public String getAddStaff(Model model){
+	public String getAddStaff(Model model,HttpSession httpSession){
+		
+		if(httpSession.getAttribute("validAdmin")==null) {
+			return "adminlogin";
+		}
+		
 		model.addAttribute("addStaffIndicator","active");
 		return "admin_add_staff";
 	}
 	
 	@GetMapping("/staffList")
-	public String getAllist(Model model) {
+	public String getAllist(Model model,HttpSession httpSession) {
+		
+		if(httpSession.getAttribute("validAdmin")==null) {
+			return "adminlogin";
+		}
+		
 		List<StaffModel> stafflist=staffService.getAllStaff();
 		model.addAttribute("staffListIndicator","active");
 		model.addAttribute("staff_list",stafflist);
@@ -134,30 +181,63 @@ public class AdminpageController {
 	}
 	
 	@GetMapping("/deleteStaff")
-	public String getDeleteStaff(@RequestParam int id) {
+	public String getDeleteStaff(@RequestParam int id,HttpSession httpSession) {
+		
+		if(httpSession.getAttribute("validAdmin")==null) {
+			return "adminlogin";
+		}
+		
 		staffService.deleteStaff(id);
 		return "redirect:/staffList";
 	}
 	
 	@GetMapping("/adminContact")
-	public String getAdminContact(@RequestParam String gmail,Model model) {
+	public String getAdminContact(@RequestParam String gmail,Model model,HttpSession httpSession) {
+		
+		if(httpSession.getAttribute("validAdmin")==null) {
+			return "adminlogin";
+		}
+		
 		model.addAttribute("gmail",gmail);
 		return "admin_contact";
 	}
 	
 	
 	@GetMapping("/viewDoctorDetails")
-	public String getadminasjdn(@RequestParam int id,Model model) {
+	public String getadminasjdn(@RequestParam int id,Model model,HttpSession httpSession) {
+		
+		if(httpSession.getAttribute("validAdmin")==null) {
+			return "adminlogin";
+		}
+		
 		DoctorModel doctorModel=doctorService.getDoctorId(id);
 		model.addAttribute("doctorObject", doctorModel);
 		return "admin_view_doctor_details";
 	}
 	
 	@GetMapping("/viewPatientDetails")
-	public String getadminViewPatient(@RequestParam int id,Model model) {
-	
+	public String getadminViewPatient(@RequestParam int id,Model model,HttpSession httpSession) {
+		
+		if(httpSession.getAttribute("validAdmin")==null) {
+			return "adminlogin";
+		}
+		
+		PatientModel patientModel=patientListApi.getPatient(id);
+	  model.addAttribute("patientObject", patientModel);
 		//model.addAttribute("doctorObject", doctorModel);
-		return "admin_view_doctor_details";
+		return "admin_view_patient_details";
+	}
+	
+	@GetMapping("/viewDoctorRequestDetails")
+	public String getadminDoctorRequest(@RequestParam int id,Model model,HttpSession httpSession) {
+		
+		if(httpSession.getAttribute("validAdmin")==null) {
+			return "adminlogin";
+		}
+		
+		DoctorModel doctorModel=doctorService.getDoctorId(id);
+		model.addAttribute("doctorObject", doctorModel);
+		return "admin_view_doctor_request_details";
 	}
 
 }
